# openapi.yaml
openapi: 3.1.0
info:
  title: 6MD Management Platform API
  description: |
    Web application API for managing Arma 3 unit operations, signups, attendance, ranks, 
    training progression, and admin workflows for the 6th Maroon Division.
  version: 1.0.0
  contact:
    name: 6th Maroon Division
  license:
    name: Proprietary
    identifier: 6MD
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://orbat.6md.net/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /orbats/{id}:
    get:
      tags:
        - ORBATs
      summary: Get ORBAT by ID
      security:
        - bearerAuth:
            - orbat:edit
      responses:
        "200":
          description: ORBAT details
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: ORBAT not found
    patch:
      tags:
        - ORBATs
      summary: Update ORBAT
      security:
        - bearerAuth:
            - orbat:edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrbatUpdateInput"
      responses:
        "200":
          description: ORBAT updated
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: ORBAT not found
    delete:
      tags:
        - ORBATs
      summary: Delete ORBAT
      security:
        - bearerAuth:
            - orbat:delete
      responses:
        "200":
          description: ORBAT deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: ORBAT not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth.js session token (cookie-based)
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: BotToken
      description: Active bot token stored in the BotToken database table. Use Authorization Bearer; environment tokens and X-BOT-API-TOKEN are not accepted.
  schemas:
...
    OrbatUpdateInput:
      type: object
      required:
        - name
        - squads
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        eventDate:
          type: string
          format: date
          nullable: true
        startTime:
          type: string
          format: time
          nullable: true
        endTime:
          type: string
          format: time
          nullable: true
        squads:
          type: array
          items:
            type: object
            required:
              - name
              - slots
            properties:
              id:
                type: integer
                nullable: true
              name:
                type: string
              orderIndex:
                type: integer
              slots:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      nullable: true
                    name:
                      type: string
                    orderIndex:
                      type: integer
                    maxSignups:
                      type: integer
                    squadRoleId:
                      type: integer
                      nullable: true
                    _deleted:
                      type: boolean
                      default: false
              _deleted:
                type: boolean
                default: false
...